+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [unreleased] 
 
 - add `can_remove` widget argument which controls whether to include a remove button
+- scroll dropdown into view when opening it
 
 ## 0.6.2 (2023-09-11)